# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Top All Money API",
        "description": "Documentation de l'API top all money",
        "contact": {
            "email": "support@tonapp.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local"
        },
        {
            "url": "https://api.tonapp.com",
            "description": "Production"
        }
    ],
    "paths": {
...
        "/api/messages": {
            "get": {
                "tags": [
                    "Messages"
                ],
                "summary": "List messages",
                "operationId": "4216f3eedb7be911361f1b91030253a8",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "isDirty",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "enum": [
                                0,
                                1
                            ]
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "example": "updated_at:desc"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 20,
                            "maximum": 200,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "post": {
                "tags": [
                    "Messages"
                ],
                "summary": "Create message",
                "operationId": "c85a9171473a13f151a8672235e01559",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MessageCreateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    }
                }
            }
        },
        "/api/messages/{id}": {
            "get": {
                "tags": [
                    "Messages"
                ],
                "summary": "Show message",
                "operationId": "2d72cf14bf2d55c9c9dfd297da04d404",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "put": {
                "tags": [
                    "Messages"
                ],
                "summary": "Update message",
                "operationId": "f5c039674f51b998b915d3172b879661",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MessageUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Messages"
                ],
                "summary": "Delete message",
                "operationId": "3900a1f5d7b34a2d1a02b42535a3fdd9",
                "responses": {
                    "200": {
                        "description": "Deleted"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Messages"
                ],
                "summary": "Partially update message",
                "operationId": "00fdc0b67fb356b142f840a26e9bf60f",
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MessageUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "MessageCreateRequest": {
                "required": [
                    "conversation_id",
                    "body"
                ],
                "properties": {
                    "conversation_id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "body": {
                        "type": "string"
                    },
                    "sender": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "MessageUpdateRequest": {
                "properties": {
                    "body": {
                        "type": "string"
                    },
                    "sender": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}